Chore: Rework for full theming
This commit removes the old, manually generated Directory Opus theme file for the Catppuccin Mocha flavor and replaces the repository with a system that allows the generation of all Catppuccin flavors in one run.

The new system works by defining Yaml files with the properties for the full XML theme file. By segmenting each configuration part based on its structure in the Directory Opus lister properties, managing and adjusting the values becomes more humanly manageable.

The Node.js-based builder generates a full XML file from all the fragmented Yaml files and packages the theme for import in Directory Opus.

It generates the following configuration parameters:
- Color Blending
- Directory Opus Colors
- Windows Colors

Additionally, the README was adjusted to better follow the Catppuccin template. It is doubtful the theme will ever become a proper port, though.